* sql query eingabe im info.php adminbereich * speicherbedarf der minimap-erzeugung stark reduziert (-> auch schneller) * minimap 2 regenerierung bei terraformern nurnoch auf knopfdruck, nicht bei jedem scrollen * menu.php wird im dummy frame nicht mehr eingebunden -> einige mysql querries weniger (hq suchen, post...) * mehrere dummy frames (10), die zyklisch benutzt werden -> beim mehrmals map-clicken geht das feedback meistens nicht verloren, sondern kommt gestaffelt
git-svn-id: svn://zwischenwelt.org:3690/zw@146 6ae2b031-6f08-0410-bc4a-c5f2e5668f8d

+++ minimap2.php
@@ -100,13 +100,14 @@ if(isset($_REQUEST["cross_x"])){
 	$clicked = true;
 }
 
+if (isset($f_regenerate_minimap2))
 if(($gUser->flags & kUserFlags_TerraFormer) || $gUser->admin){
 	for($y=0;$y<kShowSegmentHeight;++$y)
 		for($x=0;$x<kShowSegmentWidth;++$x){
 			$px = $tmp_left_seg+$x;
 			$py = $tmp_top_seg+$y;
 			$filename = "tmp/minimap/seg_".$mode."_".$px."_".$py.".png";
-			unlink($filename);
+			if (file_exists($filename)) unlink($filename);
 		}
 }
 
@@ -214,6 +215,10 @@ TileLink(4,3);
 TileLink(4,4);
 TileLink(5,5);
 ?>
+<br>
+<?php if (($gUser->flags & kUserFlags_TerraFormer) || $gUser->admin) {?>
+<a href="<?=Query("?sid=?&mode=?&cx=?&cy=?&crossx=?&crossy=?&f_regenerate_minimap2=1")?>">Terraformer : Sichtbare Minimap Felder neu generieren</a>
+<?php } // endif?>
 
 </BODY>
 </HTML>
